Memory Management Simulator - Java Swing GUI
This Project is a simulation of a Memory Manager. It includes a MVC
implemented GUI that shows you the Physical Memory and allows you to
select Processes and see their Page Tables. There is 4KB of Physical
Memory with a Page Size of 512B. This program assumes the file being
read is of the correct format.